auth_microservice/requirements.txt
bcrypt==4.2.0
certifi==2024.8.30
cryptography==43.0.3
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
tzdata==2024.2
watchdog==5.0.3
Werkzeug==3.0.4
